+++ miracle/models/__init__.py
@@ -1,8 +1,0 @@
-# Import and as a side-effect register all models here.
-
-from .base import Model  # NOQA
-from .session import (  # NOQA
-    Session,
-    URL,
-    User,
-)