+++ miracle/worker/tests.py
@@ -6,6 +6,5 @@ def test_config(celery):
     assert hasattr(celery, 'bucket')
     assert hasattr(celery, 'cache')
     assert hasattr(celery, 'crypto')
-    assert hasattr(celery, 'db')
     assert hasattr(celery, 'raven')
     assert hasattr(celery, 'stats')